moved lbfields functions to utils file
